refactor(Post): create shared type for Post
