+++ client/src/api/types/post.ts
@@ -1,7 +1,6 @@
-import { PostStatus } from '../../types/post-status'
 import { BaseModelParams, MessageResponse } from './common'
 import { BaseTagDto } from './tag'
-import { User as UserBaseDto } from '~shared/types/base'
+import { User as UserBaseDto, PostStatus } from '~shared/types/base'
 
 export type BasePostDto = BaseModelParams & {
   userId: number
