+++ client/src/pages/Post/Post.component.jsx
@@ -15,7 +15,7 @@ import {
   getPostById,
   GET_POST_BY_ID_QUERY_KEY,
 } from '../../services/PostService'
-import { PostStatus } from '../../types/post-status'
+import { PostStatus } from '~shared/types/base'
 import { TagType } from '../../types/tag-type'
 import AnswerSection from './AnswerSection/AnswerSection.component'
 import './Post.styles.scss'
