--- client/src/types/post-status.ts
@@ -1,5 +1,0 @@
-export enum PostStatus {
-  Public = 'PUBLIC',
-  Private = 'PRIVATE',
-  Archived = 'ARCHIVED',
-}
