+++ server/src/modules/post/__tests__/post.routes.spec.ts
@@ -12,7 +12,7 @@ import {
   Tag as TagModel,
   User as UserModel,
 } from '../../../models'
-import { PostStatus } from '../../../types/post-status'
+import { PostStatus } from '../../../../../shared/types/base'
 import { SortType } from '../../../types/sort-type'
 import { TagType } from '../../../types/tag-type'
 import { createTestDatabase, getModel, ModelName } from '../../../util/jest-db'
@@ -78,7 +78,7 @@ describe('/posts', () => {
     for (let title = 1; title <= 20; title++) {
       const mockPost = await Post.create({
         title: title.toString(),
-        status: PostStatus.PUBLIC,
+        status: PostStatus.Public,
         userId: mockUser.id,
       })
       mockPosts.push(mockPost)
