--- server/src/types/post-status.ts
@@ -1,5 +1,0 @@
-export const PostStatus = {
-  PUBLIC: 'PUBLIC',
-  PRIVATE: 'PRIVATE',
-  ARCHIVED: 'ARCHIVED',
-}
