+++ shared/types/base/common.ts
@@ -1,0 +1,5 @@
+import { z } from 'zod'
+
+export const BaseModel = z.object({
+  id: z.number(),
+})
