+++ shared/types/base/index.ts
@@ -1,1 +1,2 @@
+export * from './post'
 export * from './user'
